# Conflicts:
#	src/main/java/ru/tecon/admTools/systemParams/ejb/temperature/TemperatureSB.java

+++ src/main/java/ru/tecon/admTools/systemParams/ejb/temperature/TemperatureSB.java
@@ -61,7 +61,7 @@ public class TemperatureSB {
             ResultSet res = stm.executeQuery();
 
             while (res.next()) {
-                result.add(new TemperatureProp(res.getInt("x"), res.getInt("y")));
+                result.add(new TemperatureProp(res.getInt("x"), res.getDouble("y")));
             }
         } catch (SQLException e) {
             LOGGER.log(Level.WARNING, "error load temperature props", e);
@@ -124,7 +124,7 @@ public class TemperatureSB {
              CallableStatement cStm = connect.prepareCall(select)) {
             cStm.setInt(1, id);
             cStm.setInt(2, temperatureProp.getTnv());
-            cStm.setInt(3, temperatureProp.getValue());
+            cStm.setDouble(3, temperatureProp.getValue());
             cStm.setString(4, login);
             cStm.setString(5, ip);
             cStm.registerOutParameter(6, Types.SMALLINT);
